chore(common-icons): move stories to under the spec
https://github.com/akky-xxxx/project-my-homepage/issues/178
